Move WritableContents to hash package
